-add rename method -add new provider method to reservationhandler -fix some import bugs -add owner check to setBuissnesHours

+++ contracts/LockFactory.sol
@@ -1,4 +1,5 @@
-// SPDX-License-Keyentifier: MIT
+// SPDX-License-Identifier: MIT
+
 pragma solidity >=0.5.17 <0.9.0;
 pragma experimental ABIEncoderV2;
 
@@ -7,12 +8,15 @@ import "./interfaces/unlock/IUnlock.sol";
 
 import "./Owned.sol";
 
-contract LockFactory is Owned{
+contract LockFactory is Owned {
     IUnlock internal unlock;
 
-    mapping (bytes32 => IPublicLock) lockToKey;
+    mapping(bytes32 => IPublicLock) lockToKey;
 
-    function setLockAddress(address payable adr, bytes32 key) external onlyOwner {
+    function setLockAddress(address payable adr, bytes32 key)
+        external
+        onlyOwner
+    {
         lockToKey[key] = IPublicLock(adr);
     }
 
@@ -22,19 +26,20 @@ contract LockFactory is Owned{
 
     function createNewLock(bytes32 key) internal {
         //unlock.createLock(_expirationDuration, _tokenAddress, _keyPrice, _maxNumberOfKeys, _lockName, _salt);
-        IPublicLock lock = IPublicLock(0x2D7Fa4dbdF5E7bfBC87523396aFfD6d38c9520fa);
+        IPublicLock lock =
+            IPublicLock(0x2D7Fa4dbdF5E7bfBC87523396aFfD6d38c9520fa);
         lockToKey[key] = lock;
     }
 
-    function getKeyPrice(bytes32 key) external view returns (uint) {
+    function getKeyPrice(bytes32 key) external view returns (uint256) {
         return lockToKey[key].keyPrice();
     }
 
-    function updateKeyPrice(bytes32 key, uint keyPrice) external {
+    function updateKeyPrice(bytes32 key, uint256 keyPrice) external {
         lockToKey[key].updateKeyPricing(keyPrice, address(0));
     }
 
-    function getLock(bytes32 key) public view returns (IPublicLock){
+    function getLock(bytes32 key) public view returns (IPublicLock) {
         return lockToKey[key];
     }
 }
